+++ include/roq/cancel_order.h
@@ -23,8 +23,9 @@ namespace roq {
 
 //! Fields required to cancel an existing order
 struct ROQ_PUBLIC CancelOrder final {
-  std::string_view account;  //!< Account name (as known to the gateway)
-  uint32_t order_id = {};    //!< Order identifier (as known to client)
+  std::string_view account;     //!< Account name (as known to the gateway)
+  uint32_t order_id = {};       //!< Order identifier (as known to client)
+  std::string_view routing_id;  //!< Routing identifier
 };
 
 }  // namespace roq
@@ -38,10 +39,12 @@ struct fmt::formatter<roq::CancelOrder> : public roq::formatter {
         context.out(),
         R"({{)"
         R"(account="{}", )"
-        R"(order_id={})"
+        R"(order_id={}, )"
+        R"(routing_id="{}")"
         R"(}})"_fmt,
         value.account,
-        value.order_id);
+        value.order_id,
+        value.routing_id);
   }
 };
 template <>