+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes will be documented in this file.
 * `GatewaySettings.mbp_allow_remove_non_existing` to indicate if an exchange
    could possibly send updates to unknown price levels
 * `StatisticsType::FUNDING_RATE` and `StatisticsType::DAILY_FUNDING_RATE`
+* `routing_id` to `ModifyOrder` and `CancelOrder`
 
 ## 0.7.0 &ndash; 2021-04-15